refactor: extract session state helpers from AgentWorkspace
